+++ geomeny/convex-cut.hpp
@@ -1,12 +1,12 @@
 #pragma once
-#include "../template/template.hpp"
 #include "../geomeny/geomeny-template.hpp"
+#include "../template/template.hpp"
 
 int ccw_for_convexcut(const Point &a, const Point &b, const Point &c) {
-    if (cross(b-a, c-a) > EPS) return 1;
-    if (cross(b-a, c-a) < -EPS) return -1;
-    if (dot(b-a, c-a) < -EPS) return 2;
-    if (norm(b-a) < norm(c-a) - EPS) return -2;
+    if (cross(b - a, c - a) > EPS) return 1;
+    if (cross(b - a, c - a) < -EPS) return -1;
+    if (dot(b - a, c - a) < -EPS) return 2;
+    if (norm(b - a) < norm(c - a) - EPS) return -2;
     return 0;
 }
 vector<Point> crosspoint_for_convexcut(const Line &l, const Line &m) {
@@ -18,17 +18,22 @@ vector<Point> crosspoint_for_convexcut(const Line &l, const Line &m) {
 }
 vector<Point> ConvexCut(const vector<Point> &pol, const Line &l) {
     vector<Point> res;
-    for (int i = 0; i < pol.size(); ++i) {
-        Point p = pol[i], q = pol[(i+1)%pol.size()];
+    for (int i = 0; i < (int)pol.size(); ++i) {
+        Point p = pol[i], q = pol[(i + 1) % pol.size()];
         if (ccw_for_convexcut(l[0], l[1], p) != -1) {
-            if (res.size() == 0) res.push_back(p);
-            else if (!eq(p, res[res.size()-1])) res.push_back(p);
+            if (res.size() == 0)
+                res.push_back(p);
+            else if (!eq(p, res[res.size() - 1]))
+                res.push_back(p);
         }
         if (ccw_for_convexcut(l[0], l[1], p) * ccw_for_convexcut(l[0], l[1], q) < 0) {
             vector<Point> temp = crosspoint_for_convexcut(Line(p, q), l);
-            if (temp.size() == 0) continue;
-            else if (res.size() == 0) res.push_back(temp[0]);
-            else if (!eq(temp[0], res[res.size()-1])) res.push_back(temp[0]);
+            if (temp.size() == 0)
+                continue;
+            else if (res.size() == 0)
+                res.push_back(temp[0]);
+            else if (!eq(temp[0], res[res.size() - 1]))
+                res.push_back(temp[0]);
         }
     }
     return res;
