+++ test/yosupo/shortes-path.test.cpp
@@ -19,7 +19,7 @@ int main() {
         return 0;
     }
     cout << res[t] << " " << ans.size() - 1 << endl;
-    for (int i = 0; i < ans.size() - 1; i++) {
+    for (int i = 0; i < (int)ans.size() - 1; i++) {
         cout << ans[i] << " " << ans[i + 1] << endl;
     }
 }
